Bugs!!

+++ src/domains_library/domain_helpers.h
@@ -12,9 +12,9 @@ int get_num_refinement_steps_to_discretization (double side_len, double h);
 void set_benchmark_domain (struct grid *the_grid);
 void set_plain_domain (struct grid *the_grid, double sizeX, double sizeY, double sizeZ);
 
-void set_custom_mesh(struct grid *the_grid, const char *file_name, int size, bool read_fibrosis);
+void set_custom_mesh(struct grid *the_grid, const char *file_name, size_t size, bool read_fibrosis);
 
-void set_custom_mesh_with_bounds (struct grid *the_grid, const char *file_name, int size,
+void set_custom_mesh_with_bounds (struct grid *the_grid, const char *file_name, size_t size,
                                   double minx, double maxx, double miny, double maxy, double minz,
                                   double maxz,  bool read_fibrosis);
