+++ src/utils/utils.h
@@ -5,7 +5,9 @@
 #ifndef MONOALG3D_UTILS_H_H
 #define MONOALG3D_UTILS_H_H
 
-void sort_vector(double **a, int length);
+#include <stddef.h>
+
+void sort_vector(double **a, size_t length);
 int inside_mesh(double **a, double x, double y, double z, int first, int last);
 
 #endif //MONOALG3D_UTILS_H_H
